phishing_url_detector/requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
nltk==3.8.1
urllib3==2.1.0
joblib==1.3.2
Werkzeug==3.0.1
fpdf==1.7.2
selenium
webdriver-manager
beautifulsoup4
requests

flask-cors
python-whois
gunicorn
